refactor(Telemetry): Rename `analytics` to `telemetry`
