+++ lib/plugins/deploy.js
@@ -2,8 +2,8 @@
 
 const ServerlessError = require('../serverless-error');
 const cliCommandsSchema = require('../cli/commands-schema');
-const generateAnalyticsPayload = require('../utils/analytics/generatePayload');
-const { report: analyticsReport } = require('../utils/analytics');
+const generateTelemetryPayload = require('../utils/telemetry/generatePayload');
+const { report: telemetryReport } = require('../utils/telemetry');
 const processBackendNotificationRequest = require('../utils/processBackendNotificationRequest');
 
 class Deploy {
@@ -53,8 +53,8 @@ class Deploy {
             await this.serverless.pluginManager.spawn('package');
           }
         } finally {
-          this.deferredBackendNotificationRequest = analyticsReport(
-            Object.assign(await generateAnalyticsPayload(this.serverless), { command: 'deploy' })
+          this.deferredBackendNotificationRequest = telemetryReport(
+            Object.assign(await generateTelemetryPayload(this.serverless), { command: 'deploy' })
           );
         }
       },
