+++ lib/utils/telemetry/areDisabled.js
