+++ lib/utils/telemetry/cache-path.js
