+++ lib/utils/telemetry/generatePayload.js
