+++ lib/utils/telemetry/index.js
@@ -6,9 +6,9 @@ const ensurePlainObject = require('type/plain-object/ensure');
 const { v1: uuid } = require('uuid');
 const fetch = require('node-fetch');
 const fse = require('fs-extra');
-const analyticsUrl = require('@serverless/utils/analytics-and-notfications-url');
+const telemetryUrl = require('@serverless/utils/analytics-and-notfications-url');
 const log = require('../log/serverlessLog');
-const areAnalyticsDisabled = require('./areDisabled');
+const isTelemetryDisabled = require('./areDisabled');
 const cacheDirPath = require('./cache-path');
 
 const timestampWeekBefore = Date.now() - 1000 * 60 * 60 * 24 * 7;
@@ -54,7 +54,7 @@ async function request(payload, { ids, timeout } = {}) {
   let response;
   const body = JSON.stringify(payload);
   try {
-    response = await fetch(analyticsUrl, {
+    response = await fetch(telemetryUrl, {
       headers: {
         'content-type': 'application/json',
       },
@@ -91,9 +91,9 @@ async function request(payload, { ids, timeout } = {}) {
 
 async function report(payload, options = {}) {
   ensurePlainObject(payload);
-  if (!analyticsUrl) return null;
+  if (!telemetryUrl) return null;
   const isForced = options && options.isForced;
-  if (areAnalyticsDisabled && !isForced) return null;
+  if (isTelemetryDisabled && !isForced) return null;
   if (!cacheDirPath) return request([payload]);
   const id = uuid();
 
@@ -125,9 +125,9 @@ async function sendPending(options = {}) {
   if (options.serverlessExecutionSpan) {
     options.serverlessExecutionSpan.then(markServerlessRunEnd, markServerlessRunEnd);
   }
-  if (areAnalyticsDisabled && !isForced) return null;
+  if (isTelemetryDisabled && !isForced) return null;
   if (!cacheDirPath) return null;
-  if (!analyticsUrl) return null;
+  if (!telemetryUrl) return null;
   let dirFilenames;
   try {
     dirFilenames = await fse.readdir(cacheDirPath);
