+++ package.json
@@ -75,7 +75,7 @@
   "devDependencies": {
     "@commitlint/cli": "^12.1.1",
     "@serverless/eslint-config": "^3.0.0",
-    "@serverless/test": "^8.0.0",
+    "@serverless/test": "^8.0.2",
     "chai": "^4.3.4",
     "chai-as-promised": "^7.1.1",
     "cos-nodejs-sdk-v5": "^2.9.13",
